refactor(server): move handleBind + runExtendedQueryMessage into conn_extended_query.go (#722 phase 1.5)
Finishes the extended-query consolidation started in #728, which split
server/conn.go into conn_*.go files but deliberately left handleBind and
runExtendedQueryMessage in conn.go while #717/#718/#720 were editing them
concurrently. Those PRs are now merged, so this is a pure mechanical move:

- handleBind moved from conn.go to conn_extended_query.go (alongside
  handleParse/handleDescribe/handleExecute/handleClose/readCString).
- runExtendedQueryMessage (the extended-query dispatch + skip-until-Sync
  arming) moved to conn_extended_query.go.
- conn.go drops its now-unused encoding/binary import; conn_extended_query.go
  gains io for handleBind's io.ReadFull.

The ignoreTillSync / errorResponsesSent fields stay on clientConn in conn.go.
messageLoop, handleStartup, handleQuery, and isDMLReturning (+ helpers) remain
in conn.go per the DML RETURNING and dispatch invariants. Zero renames, zero
signature changes, zero logic edits, no behavior change.
